requirements.txt
pydantic>=2.0.0
python-dotenv>=1.0.0
openai>=1.0.0
hello-agents>=0.1.0
fastapi>=0.110.0
uvicorn>=0.27.0
pymupdf>=1.24.0
rank_bm25>=0.2.2
jieba>=0.42.1
qdrant-client>=1.7.0
dashscope>=1.14.0
transformers>=4.36.0
torch>=2.0.0
